+++ src/backend/ImageBackend.cc
@@ -64,10 +64,9 @@ void ImageBackend::Initialize(Local<Object> target) {
 	ImageBackend::constructor.Reset(ctor);
 	ctor->InstanceTemplate()->SetInternalFieldCount(1);
 	ctor->SetClassName(Nan::New<String>("ImageBackend").ToLocalChecked());
-  Local<Context> v8ctx = Nan::GetCurrentContext();
-	target->Set(v8ctx,
-              Nan::New<String>("ImageBackend").ToLocalChecked(),
-              ctor->GetFunction(v8ctx).ToLocalChecked());
+  Nan::Set(target,
+           Nan::New<String>("ImageBackend").ToLocalChecked(),
+           Nan::GetFunction(ctor).ToLocalChecked()).Check();
 }
 
 NAN_METHOD(ImageBackend::New) {