+++ src/backend/PdfBackend.cc
@@ -43,10 +43,9 @@ void PdfBackend::Initialize(Local<Object> target) {
   PdfBackend::constructor.Reset(ctor);
   ctor->InstanceTemplate()->SetInternalFieldCount(1);
   ctor->SetClassName(Nan::New<String>("PdfBackend").ToLocalChecked());
-  Local<Context> v8ctx = Nan::GetCurrentContext();
-  target->Set(v8ctx,
-              Nan::New<String>("PdfBackend").ToLocalChecked(),
-              ctor->GetFunction(v8ctx).ToLocalChecked());
+  Nan::Set(target,
+           Nan::New<String>("PdfBackend").ToLocalChecked(),
+           Nan::GetFunction(ctor).ToLocalChecked()).Check();
 }
 
 NAN_METHOD(PdfBackend::New) {