+++ src/backend/SvgBackend.cc
@@ -51,10 +51,9 @@ void SvgBackend::Initialize(Local<Object> target) {
   SvgBackend::constructor.Reset(ctor);
   ctor->InstanceTemplate()->SetInternalFieldCount(1);
   ctor->SetClassName(Nan::New<String>("SvgBackend").ToLocalChecked());
-  Local<Context> v8ctx = Nan::GetCurrentContext();
-  target->Set(v8ctx,
-              Nan::New<String>("SvgBackend").ToLocalChecked(),
-              ctor->GetFunction(v8ctx).ToLocalChecked());
+  Nan::Set(target,
+           Nan::New<String>("SvgBackend").ToLocalChecked(),
+           Nan::GetFunction(ctor).ToLocalChecked()).Check();
 }
 
 NAN_METHOD(SvgBackend::New) {